+++ client-side/src/app/portals/Bam/components/calendar/edit-batch/edit-batch.component.html
