+++ client-side/src/app/portals/Bam/components/calendar/remove-associate-from-batch/remove-associate-from-batch.component.css
