+++ client-side/src/app/portals/Bam/components/calendar/view-associates/view-associates.component.css
