+++ client-side/src/app/portals/Bam/components/calendar/view-associates/view-associates.component.ts
@@ -1,23 +1,24 @@
 import { Component, OnInit } from '@angular/core';
-import { UsersService } from '../../services/users.service';
-import { BamUser } from '../../models/bamuser.model';
-import { SearchPipe } from '../../pipes/search.pipe';
 import { NgxPaginationModule } from 'ngx-pagination';
 
-
-
+import { UsersService } from '../../../services/users.service';
+import { BamUser } from '../../../models/bamuser.model';
+import { SearchPipe } from '../../../pipes/search.pipe';
 
 @Component({
   selector: 'app-view-associates',
   templateUrl: './view-associates.component.html',
   styleUrls: ['./view-associates.component.css']
 })
+
 export class ViewAssociatesComponent implements OnInit {
   p: number = 1;
   associateList: BamUser[];
   searchTerm: string;
   order: string;
+
   constructor(private usersService: UsersService) { }
+
   ngOnInit() {
     this.loadAssociatesInBatch();
   }
